Downgrade Analyzer dependency
Previously we were depending on "analyzer" 0.39.4+, however this was a version greater than the one Flutter's stable is using. This causes this library to not work on Flutter's Stable branch which is undesirable (and at this point unnecessary). As there was no real reason for this enforcing, I'm downgrading to match the version on Flutter stable.