Changes to explain better a usage of nuvigator
With these changes, novice users can better understand how to use nuvigator from scratch. For those who have never used it before, the example "sample" can become a bit confusing because it does not correctly explain the import part of the generated file and how to generate this file. If we explain just at the end of README.md how to use build_runner, the rest of the information ends up shuffling the understanding of how to create a simple route scheme. As it is now, beginners and advanced users can understand, thus, making the documentation more accessible.